// WADNR.Web/src/app/shared/generated/model/project-update-configuration-detail.ts
export class ProjectUpdateConfigurationDetail { 
    ProjectUpdateConfigurationID?: number;
    EnableProjectUpdateReminders?: boolean;
    ProjectUpdateKickOffDate?: string | null;
    ProjectUpdateKickOffIntroContent?: string | null;
    SendPeriodicReminders?: boolean;
    ProjectUpdateReminderInterval?: number | null;
    ProjectUpdateReminderIntroContent?: string | null;
    SendCloseOutNotification?: boolean;
    ProjectUpdateCloseOutDate?: string | null;
    ProjectUpdateCloseOutIntroContent?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}