// WADNR.Web/src/app/shared/generated/model/project-update-configuration-upsert-request.ts
export class ProjectUpdateConfigurationUpsertRequest { 
    EnableProjectUpdateReminders: boolean;
    ProjectUpdateKickOffDate?: string | null;
    ProjectUpdateKickOffIntroContent?: string | null;
    SendPeriodicReminders: boolean;
    ProjectUpdateReminderInterval?: number | null;
    ProjectUpdateReminderIntroContent?: string | null;
    SendCloseOutNotification: boolean;
    ProjectUpdateCloseOutDate?: string | null;
    ProjectUpdateCloseOutIntroContent?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}